% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easySNA.R
\name{actionSNA}
\alias{actionSNA}
\title{actionSNA}
\usage{
actionSNA(docs)
}
\arguments{
\item{docs}{dataframe}
}
\value{
matrix
}
\description{
Analyse SNA by action
}
\examples{
actionSNA(docs = docs)
}
