% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easySNA.R
\name{fromtoSNA}
\alias{fromtoSNA}
\title{fromtoSNA}
\usage{
fromtoSNA(docs)
}
\arguments{
\item{docs}{dataframe. It has from, to, cnt column.}
}
\value{
matrix
}
\description{
Analyse SNA for from to data
}
\examples{
fromtoSNA(docs = docs)
}
