% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easySNA.R
\name{keywordsSNA}
\alias{keywordsSNA}
\title{keywordsSNA}
\usage{
keywordsSNA(docs, support = 0.1, confidence = 0.8)
}
\arguments{
\item{docs}{vector, split by enter}

\item{support}{minimum support}

\item{confidence}{minimum confidence}
}
\value{
rules
}
\description{
Analyse SNA by keywords
}
\examples{
keywordsSNA(docs = docs, support=0.1, confidence=0.8)
}
